Integrate Ranked Patient Match with Compound Key
Add transitionary class CompoundKeyRankedPatientMatch that goes in place
of RankedPatientMatch which includes the new kit id to map to. Then
these objects are used to create a list of mapped matches which are used
to create a list of RankedPatientItems that are displayed in the UI
matches.
The new class shouldn't be a long term solution and is just to get the
feature under test as it just takes all the attributes from
rankedpatientmatch and patient and puts them in a class to map the sql
database too.

ZZS-96